✨ `WallpaperImageFetcher` to fetch wallpaper image
