+++ Loop/Config.xcconfig
@@ -5,5 +5,5 @@
 //  Created by Kai Azim on 2023-08-09.
 //
 
-VERSION = 1.2.0
-BUILD_NUMBER = 1546
+VERSION = 0.0.0
+BUILD_NUMBER = 0
